Updated to v1.1
A couple changes I've made: First I removed the update command, because it requires the update script to exist, and it won't after the make script is run. The next thing is I added a --no-commit option to the push command. I refactored the code so that there's a function for every git command.